# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TARA Travel Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a single trip with full detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{id}/boarding-pass:
    get:
      operationId: getBoardingPass
      tags: [trips]
      summary: Get boarding pass for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Boarding pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardingPass"
  /trips/{id}/alerts:
    get:
      operationId: getTripAlerts
      tags: [trips]
      summary: Get alerts for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripAlert"

  # ─── AGENTS ───────────────────────────────────────────────────────────────
...
  /wallet/fund:
    post:
      operationId: fundWallet
      tags: [wallet]
      summary: Add funds to wallet (simulated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundInput"
      responses:
        "200":
          description: Funded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  # ─── ITINERARIES ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        seatPreference:
          type: ["string", "null"]
        mealPreference:
          type: ["string", "null"]
        airlinePreference:
          type: ["string", "null"]
        walletBalance:
          type: number
        taraPoints:
          type: integer
        onboardingComplete:
          type: boolean
        plan:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - walletBalance
        - taraPoints
        - onboardingComplete
        - plan
        - createdAt
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        seatPreference:
          type: string
        mealPreference:
          type: string
        airlinePreference:
          type: string
    OnboardingInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: ["string", "null"]
        seatPreference:
          type: string
        mealPreference:
          type: string
        airlinePreference:
          type: string
        whatsappNumber:
          type: string
      required:
        - name
        - phone

    # ─── DASHBOARD ─────────────────────────────────────────────────────────
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        flightNumber:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        gate:
          type: ["string", "null"]
        terminal:
          type: ["string", "null"]
        seat:
          type: ["string", "null"]
        flightStatus:
          type: string
        hotelName:
          type: ["string", "null"]
        hotelCheckIn:
          type: ["string", "null"]
        hotelCheckOut:
          type: ["string", "null"]
        budgetNgn:
          type: ["integer", "null"]
        spentNgn:
          type: ["integer", "null"]
        pricePaidNgn:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - status
        - origin
        - originCode
        - destination
        - destinationCode
        - airline
        - airlineCode
        - flightNumber
        - departureTime
        - arrivalTime
        - flightStatus
        - pricePaidNgn
        - createdAt
    TripDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        flightNumber:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        gate:
          type: ["string", "null"]
        terminal:
          type: ["string", "null"]
        seat:
          type: ["string", "null"]
        flightStatus:
          type: string
        hotelName:
          type: ["string", "null"]
        hotelAddress:
          type: ["string", "null"]
        hotelCheckIn:
          type: ["string", "null"]
        hotelCheckOut:
          type: ["string", "null"]
        budgetNgn:
          type: ["integer", "null"]
        spentNgn:
          type: ["integer", "null"]
        pricePaidNgn:
          type: integer
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/TripAlert"
        agentLogs:
          type: array
          items:
            $ref: "#/components/schemas/AgentLog"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - status
        - origin
        - originCode
        - destination
        - destinationCode
        - airline
        - airlineCode
        - flightNumber
        - departureTime
        - arrivalTime
        - flightStatus
        - pricePaidNgn
        - alerts
        - agentLogs
        - createdAt
    TripInput:
      type: object
      properties:
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        flightNumber:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        seat:
          type: string
        hotelName:
          type: string
        hotelCheckIn:
          type: string
        hotelCheckOut:
          type: string
        budgetNgn:
          type: integer
        pricePaidNgn:
          type: integer
      required:
        - origin
        - originCode
        - destination
        - destinationCode
        - airline
        - airlineCode
        - flightNumber
        - departureTime
        - arrivalTime
        - pricePaidNgn
    TripUpdate:
      type: object
      properties:
        status:
          type: string
        gate:
          type: string
        terminal:
          type: string
        flightStatus:
          type: string
        hotelName:
          type: string
        hotelCheckIn:
          type: string
        hotelCheckOut:
          type: string
        spentNgn:
          type: integer
    BoardingPass:
      type: object
      properties:
        tripId:
          type: integer
        passengerName:
          type: string
        flightNumber:
          type: string
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        departureTime:
          type: string
          format: date-time
        gate:
          type: ["string", "null"]
        terminal:
          type: ["string", "null"]
        seat:
          type: ["string", "null"]
        airline:
          type: string
        qrData:
          type: string
        version:
          type: integer
        lastUpdated:
          type: string
          format: date-time
      required:
        - tripId
        - passengerName
        - flightNumber
        - origin
        - originCode
        - destination
        - destinationCode
        - departureTime
        - airline
        - qrData
        - version
        - lastUpdated
    TripAlert:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        type:
          type: string
        message:
          type: string
        severity:
          type: string
        agentName:
          type: string
        dismissed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - type
        - message
        - severity
        - agentName
        - dismissed
        - createdAt

    # ─── AGENTS ────────────────────────────────────────────────────────────
...
    AgentLog:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: string
        agentName:
          type: string
        tripId:
          type: ["integer", "null"]
        userId:
          type: integer
        action:
          type: string
        result:
          type: string
        severity:
          type: string
        notified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - agentName
        - userId
        - action
        - result
        - severity
        - notified
        - createdAt

    # ─── NOTIFICATIONS ─────────────────────────────────────────────────────
...
    Wallet:
      type: object
      properties:
        balance:
          type: number
        taraPoints:
          type: integer
        totalSaved:
          type: number
        totalSpent:
          type: number
      required:
        - balance
        - taraPoints
        - totalSaved
        - totalSpent
...
    FundInput:
      type: object
      properties:
        amountNgn:
          type: number
      required:
        - amountNgn

    # ─── ITINERARIES ───────────────────────────────────────────────────────
...