# lib/api-zod/src/generated/types/boardingPass.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface BoardingPass {
  tripId: number;
  passengerName: string;
  flightNumber: string;
  origin: string;
  originCode: string;
  destination: string;
  destinationCode: string;
  departureTime: Date;
  /** @nullable */
  gate?: string | null;
  /** @nullable */
  terminal?: string | null;
  /** @nullable */
  seat?: string | null;
  airline: string;
  qrData: string;
  version: number;
  lastUpdated: Date;
}