# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  name?: string;
  phone?: string;
  photoUrl?: string;
  seatPreference?: string;
  mealPreference?: string;
  airlinePreference?: string;
}