# lib/api-zod/src/generated/types/tripAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TripAlert {
  id: number;
  tripId: number;
  type: string;
  message: string;
  severity: string;
  agentName: string;
  dismissed: boolean;
  createdAt: Date;
}