# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TARA Travel Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  status?: string;
  gate?: string;
  terminal?: string;
  flightStatus?: string;
  hotelName?: string;
  hotelCheckIn?: string;
  hotelCheckOut?: string;
  spentNgn?: number;
}